even more fixes
