+++ openssl-sys/src/handwritten/mod.rs
@@ -13,7 +13,9 @@ pub use handwritten::ocsp::*;
 pub use handwritten::pkcs7::*;
 pub use handwritten::rsa::*;
 pub use handwritten::sha::*;
+pub use handwritten::ssl::*;
 pub use handwritten::stack::*;
+pub use handwritten::tls1::*;
 pub use handwritten::types::*;
 pub use handwritten::x509::*;
 pub use handwritten::x509_vfy::*;
@@ -34,7 +36,9 @@ mod ocsp;
 mod pkcs7;
 mod rsa;
 mod sha;
+mod ssl;
 mod stack;
+mod tls1;
 mod types;
 mod x509;
 mod x509_vfy;
